+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `JNIEnv#define_unnamed_class` function that allows loading a class without
+  specifying its name. (#246)
+
 ## [0.17.0] — 2020-06-30
 
 ### Added